debug chan set 1204_005(add rssi from txs)

+++ bsp/kernel/bcm_v5_04L_02p1_4.19/bcmdrivers/broadcom/net/wl/impl87/main/src/wl/sys/wlc_qq_struct.h
@@ -113,7 +113,7 @@ struct phy_info_qq {
     uint8 BW[RATESEL_MFBR_NUM];
     uint32 ISSGI[RATESEL_MFBR_NUM];
     int8 SNR;
-    int8 RSSI;
+    int32 RSSI;
     int8 RSSI_loc;//0:wlc_cfp.c;1:wlc_qq.c;2:wlc_rx;3.wlc_lq_rssi_get;4.cfp...!= FC_TYPE_DATA
     int16 RSSI_type;//数据包类型
     int16 RSSI_subtype;//数据包类型
